Bug #1942152 update to 3.14 & compitable with k515
[root cause  ] update to 3.14 & compitable with k515
[changes     ] update to 3.14 & compitable with k515
[side effects] None
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:n6pro
[test case]:boot
[reviewers   ] wenming.wu
[change_type ] feature
[tag_product ] common

+++ img_mem/img_mem_anonymous.c
@@ -65,7 +65,7 @@ struct buffer_data {
 };
 
 static int anonymous_heap_import(struct device *device, struct heap *heap,
-						size_t size, enum img_mem_attr attr, uint64_t buf_fd,
+						size_t size, enum img_mem_attr attr, int buf_fd,
 						struct page **pages, struct buffer *buffer)
 {
 	struct buffer_data *data;
